+++ TestToDo.Infrastructure/Repositories/ToDoItemRepository.cs
@@ -14,15 +14,15 @@ public class ToDoItemRepository : IToDoItemRepository
         _context = context ?? throw new ArgumentNullException(nameof(context));
     }
     
-    public async Task<IReadOnlyCollection<ToDoItem>> GetToDoItemsAsync(CancellationToken cancellationToken)
+    public async Task<IReadOnlyCollection<ToDoItem>> GetToDoItemsAsync(Guid userId, CancellationToken cancellationToken)
     {
-        return await _context.ToDoItems.AsNoTracking().ToListAsync(cancellationToken);
+        return await _context.ToDoItems.AsNoTracking().Where(i=>i.UserId == userId).ToListAsync(cancellationToken);
     }
 
-    public async Task<ToDoItem?> GetToDoItemByIdAsync(Guid id, CancellationToken cancellationToken)
+    public async Task<ToDoItem?> GetToDoItemByIdAsync(Guid id, Guid userId,  CancellationToken cancellationToken)
     {
         return await
-            _context.ToDoItems.Where(i => i.Id == id).FirstOrDefaultAsync(cancellationToken);
+            _context.ToDoItems.Where(i => i.Id == id && i.UserId == userId).FirstOrDefaultAsync(cancellationToken);
     }
     public async Task AddToDoItemAsync(ToDoItem toDoItem, CancellationToken cancellationToken)
     {
@@ -30,9 +30,9 @@ public class ToDoItemRepository : IToDoItemRepository
         await _context.ToDoItems.AddAsync(toDoItem, cancellationToken);
     }
     
-    public async Task DeleteToDoItemAsync(Guid id, CancellationToken cancellationToken)
+    public async Task DeleteToDoItemAsync(Guid id, Guid userId, CancellationToken cancellationToken)
     {
-        await _context.ToDoItems.Where(i => i.Id == id).ExecuteDeleteAsync(cancellationToken);
+        await _context.ToDoItems.Where(i => i.Id == id && i.UserId == userId).ExecuteDeleteAsync(cancellationToken);
     }
 
     public async Task SaveChangesAsync(CancellationToken cancellationToken)
